feat: cria tipos do typescript
