+++ src/containers/App.tsx
@@ -1,11 +1,12 @@
-import { Box, Container } from '@mui/material'
+import { useSelector } from 'react-redux'
 
+import { Box, Container } from '@mui/material'
 import Header from '@/components/Header'
 import Guide from '@/components/Guide'
-import UploadArea from '@/components/UploadArea'
-import { useSelector } from 'react-redux'
-import { selectAcademicHistory } from '@/store/recommendation'
 import RecommendationOptions from '@/components/RecommendationOptions'
+import AcademicHistoryUpload from '@/components/AcademicHistoryUpload'
+
+import { selectAcademicHistory } from '@/store/recommendation'
 
 function App() {
   const academicHistory = useSelector(selectAcademicHistory)
@@ -16,7 +17,11 @@ function App() {
 
       <Box py={5}>
         <Container maxWidth="md">
-          {academicHistory ? <RecommendationOptions /> : <UploadArea />}
+          {academicHistory ? (
+            <RecommendationOptions />
+          ) : (
+            <AcademicHistoryUpload />
+          )}
 
           <Guide />
         </Container>
