+++ src/store/recommendation.ts
@@ -1,27 +1,6 @@
 /* eslint-disable no-param-reassign */
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
-import { RootState } from '@/store'
-
-interface AcademicHistoryWorkload {
-  required: number
-  integrated: number
-  pending: number
-}
-
-interface AcademicHistory {
-  curriculumId: string
-  approvedCourses: string[]
-  workload: {
-    mandatory: AcademicHistoryWorkload
-    optional: AcademicHistoryWorkload
-    supplementary: AcademicHistoryWorkload
-    total: AcademicHistoryWorkload
-  }
-}
-
-interface RecommendationStore {
-  academicHistory: AcademicHistory | null
-}
+import { AcademicHistory, RecommendationStore, RootState } from '@/types'
 
 const initialState: RecommendationStore = {
   academicHistory: null,
@@ -31,7 +10,7 @@ export const recommendationSlice = createSlice({
   name: 'recommendation',
   initialState,
   reducers: {
-    setAcademicHistoryData: (state, action) => {
+    setAcademicHistoryData: (state, action: PayloadAction<AcademicHistory>) => {
       state.academicHistory = action.payload
     },
   },
